Rename runtime-manager to koord-runtime-proxy
