+++ .github/workflows/ci.yaml
@@ -86,7 +86,7 @@ jobs:
           make vet && git add pkg cmd &&
           git diff --cached --exit-code || (echo 'Please run "make vet" to verify govet' && exit 1);
       - name: Run Go build
-        run: make build-koordlet build-koord-manager build-koord-scheduler build-runtime-manager
+        run: make build-koordlet build-koord-manager build-koord-scheduler build-koord-runtime-proxy
       - name: Run Go test
         run: make test
       - name: Upload coverage to Codecov
